Tests/LibWeb/Text/input/HTML/parser-streams-crlf-split.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    asyncTest(async (done) => {
        internals.setTestTimeout(5000);

        const httpServer = httpTestServer();
        const prefix = `<!DOCTYPE html><body><span id="target">line1\r`;
        const body = `${prefix}\nline2</span><script>window.onload = () => parent.postMessage(document.getElementById("target").textContent, "*");<\/script>`;
        const url = await httpServer.createEcho("GET", "/parser-streams-crlf-split", {
            status: 200,
            headers: {
                "Content-Type": "text/html",
            },
            body,
        });

        const frame = document.createElement("iframe");
        window.addEventListener("message", event => {
            const text = event.data;
            println(JSON.stringify(text));
            println(`length: ${text.length}`);
            done();
        }, { once: true });

        frame.src = `${url}?chunks=${new TextEncoder().encode(prefix).length}&chunk_delay_ms=100`;
        document.body.appendChild(frame);
    });
</script>
