Tests/LibWeb/Text/expected/HTML/parser-streams-crlf-split.txt

"line1\nline2"
length: 11
